requirements-dev.txt
pytest>=7.4.0
pytest-homeassistant-custom-component>=0.13.0
pyserial>=3.5
ruff>=0.3.0
